fix bugs save parameter to a csv file change PWParametor to PupilwareCore add more missing icons

+++ Pupilware/PupilwareCore/Algorithm/BlinkDetection.cpp
@@ -58,7 +58,7 @@ namespace pw {
 
             if( maxSize < (int)contours[i].size() )
             {
-                maxSize = contours[i].size();
+                maxSize = (int)contours[i].size();
                 bigBlobIndex = i;
             }
         }
